std/var: Add support for floating point correction (#4674)
* std/var: Add support for floating point correction

See pytorch/pytorch#94073

* Remove .torch_pin

---------

+++ torch_xla/csrc/ops/std_mean.cpp
@@ -10,7 +10,7 @@ namespace {
 
 xla::Shape NodeOutputShape(const torch::lazy::Value& input,
                            std::vector<int64_t>& dimensions,
-                           bool keep_reduced_dimensions, int64_t correction) {
+                           bool keep_reduced_dimensions, double correction) {
   auto lower_for_shape_fn_std_mean =
       [&](absl::Span<const xla::XlaOp> operands) -> xla::XlaOp {
     xla::XlaOp std = BuildStdDeviation(operands[0], dimensions,
@@ -25,7 +25,7 @@ xla::Shape NodeOutputShape(const torch::lazy::Value& input,
 }  // namespace
 
 StdMean::StdMean(const torch::lazy::Value& input,
-                 std::vector<int64_t> dimensions, int64_t correction,
+                 std::vector<int64_t> dimensions, double correction,
                  bool keep_reduced_dimensions)
     : XlaNode(
           torch::lazy::OpKind(at::aten::std_mean), {input},
